CentOS 8: Working KS for semi-automatic installation of servers
Attention: There seems to be a bug in CentOS 8 (including 8.1.1911)
preventing automatic loading of ks.cfg from OEMDRV to work (at least it did
not in our tests). You can add inst.ks=hd:sdX1/ks.cfg to he setup command
line as workaround. This is not	an error of this kickstart file.

This commit contains also Smaller updates in README and CentOS 7 kickstart
file